requirements.txt
aiocached==0.3
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.4
fastapi-cli==0.0.5
fastapi-mail==1.4.1
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.10
pillow==11.0.0
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.16
PyYAML==6.0.2
requests==2.32.3
rich==13.9.3
shellingham==1.5.4
sniffio==1.3.1
starlette==0.41.2
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
